refactor: localize conversion helper logic
